google_api/requirements.txt
django==4.1.7
django-apscheduler==0.6.2
google-api-python-client==2.82.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
psycopg2-binary==2.9.5
python-telegram-bot==20.1
djangorestframework==3.14.0
django-cors-headers==3.14.0
